task 6

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int II;
 	float f;
 
-	printf("size of a char: %lu byte(s)\n", sizeof(c));
-	printf("size of a int: %lu byte(s)\n", sizeof(i));
-	printf("size of a long int: %lu byte(s)\n", sizeof(I));
-	printf("size of a long long int: %lu byte(s)\n", sizeof(II));
-	printf("size of a float: %lu byte(s)\n", sizeof(f));
+	printf("size of a char: %lu byte(s)\n", (unsigned long) sizeof(c));
+	printf("size of a int: %lu byte(s)\n", (unsigned long) sizeof(i));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long) sizeof(I));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(II));
+	printf("size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
 	return (0);
 }
